backend/requirements.txt
asgiref==3.4.1
certifi==2021.10.8
cffi==1.14.6
charset-normalizer==2.0.7
cryptography==35.0.0
defusedxml==0.7.1
Django==3.2.8
django-allauth==0.45.0
django-cors-headers==3.10.0
django-debug-toolbar==3.2.2
django-rest-auth==0.9.5
djangorestframework==3.12.4
idna==3.2
oauthlib==3.1.1
psycopg2==2.9.1
psycopg2-binary==2.9.1
pycparser==2.20
PyJWT==2.2.0
python-magic-bin==0.4.14
python3-openid==3.2.0
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.7
whitenoise==5.2.0
coverage